Move to Headers::Normalizer
